+++ include/libopencm3/stm32/common/flash_common_f234.h
@@ -74,8 +74,6 @@
 
 BEGIN_DECLS
 
-void flash_unlock(void);
-void flash_lock(void);
 void flash_clear_pgperr_flag(void);
 void flash_clear_eop_flag(void);
 void flash_clear_status_flags(void);
