+++ include/libopencm3/stm32/l4/flash.h
@@ -224,8 +224,6 @@
 
 BEGIN_DECLS
 
-void flash_unlock(void);
-void flash_lock(void);
 void flash_clear_pgperr_flag(void);
 void flash_clear_eop_flag(void);
 void flash_wait_for_last_operation(void);
