+++ lib/stm32/l4/Makefile
@@ -47,7 +47,7 @@ OBJS            += exti_common_all.o
 OBJS            += adc_common_v2.o adc_common_v2_multi.o
 OBJS            += crc_common_all.o crc_v2.o
 OBJS            += crs_common_all.o
-OBJS            += flash_common_all.o
+OBJS            += flash_common_all.o flash_common_f.o
 OBJS            += rng_common_v1.o
 OBJS            += timer_common_all.o
 OBJS            += i2c_common_v2.o
